feat: add monitorManager to support fork and procfile mode
