+++ packages/pandora/src/daemon/spawnWrapper.ts
@@ -1,33 +1,11 @@
 'use strict';
-import {GlobalConfigProcessor} from '../universal/GlobalConfigProcessor';
+
 const wrap = require('spawn-wrap');
-const hook = require('module-hook');
-const shimmer = require('shimmer');
-import {resolve} from 'path';
+import {MonitorManager} from '../monitor/MonitorManager';
 
-const globalConfigProcessor = GlobalConfigProcessor.getInstance();
-const globalConfig = globalConfigProcessor.getAllProperties();
-const hooks = globalConfig['hooks'];
-
-/**
- * hooks: {
- *   logger: Hooks.logger
- * }
- */
-for(const hookName in hooks) {
-  if(hooks[hookName]) {
-    try {
-      let module = hooks[hookName];
-      const m = typeof module === 'string' ? require(resolve(module)) : module;
-      m(hook, shimmer);
-      console.log(`${hookName} hook enabled`);
-    } catch (err) {
-      console.log(err);
-      console.log(`enable ${hookName} hook went wrong`);
-    }
-  } else {
-    console.log(`${hookName} hook disabled`);
-  }
+try {
+  MonitorManager.injectProcessMonitor();
+} catch (err) {
+  console.log(err);
 }
-
 wrap.runMain();
